converted to ecma script project, much more usable....
