+++ app/js/TheatreActor.js
@@ -1,4 +1,4 @@
-class TheatreActor {
+export class TheatreActor {
     constructor(actor, navElement) {
         this.actor = actor;
         this.navElement = navElement;
